Add files via upload
This code is intended for knitters, to compare different types of yarn for the Monday Sweater by PetitKnit. It is possible to insert a primary and secondary yarn. 
The code will print the total yardage needed for the size(s) chosen, as well as the number of skeins needed and the total cost. 
It also provides extra information about surplus yarn in both percentage and yardage (in meters) of the last skein.